feat: make unit's lat/lng float

+++ database/migrations/2022_01_31_181542_create_units_table.php
@@ -16,8 +16,8 @@ class CreateUnitsTable extends Migration
         Schema::create('units', function (Blueprint $table) {
             $table->id();
             $table->string('status');
-            $table->double('latitude')->nullable();
-            $table->double('longitude')->nullable();
+            $table->float('latitude')->nullable();
+            $table->float('longitude')->nullable();
             $table->string('phone_number')->unique();
             $table->timestamps();
         });
